Call HA054841:
legislation_xhtml_core_vanilla.xslt amended so that the condition for including parent ListItem number (which must be required somewhere to be there in the first place) has an additional check that the current ListItem does not have a NumberOverride attribute. If it does then the parent number is not output. (10120)